Fix unstable sceanrio test for CallbackDebugBehavior.
