+++ src/System.Private.ServiceModel/tools/IISHostedWcfService/App_code/IWcfDuplexService.cs
@@ -105,9 +105,6 @@ namespace WcfService
     {
         [OperationContract]
         string Hello(string greeting, bool includeExceptionDetailInFaults);
-
-        [OperationContract]
-        bool GetResult(bool includeExceptionDetailInFaults);
     }
 
     public interface IWcfDuplexService_CallbackDebugBehavior_Callback
